Added the buffered_stream feature to hide the BufferedStream for stable users
